requirements.txt
asgiref==3.2.3
autopep8==1.5
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
Markdown==3.1.1
oauthlib==3.1.0
Pillow==7.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
python-decouple==3.3
python-dotenv==0.10.5
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
